mqtt_display_mouse/python_gui/mqtt_mouse_event_client/requirements.txt
paho-mqtt==1.5.1
Pillow>=9.0.1
